Rename filenames to BayerContourNext
