+++ lib/core/device.js
@@ -31,7 +31,7 @@ var abbottFreeStyle = require('../drivers/abbottFreeStyle');
 var asanteSnap = require('../drivers/asanteDriver');
 var insuletOmniPod = require('../drivers/insuletDriver');
 var oneTouchUltra2 = require('../drivers/oneTouchUltra2');
-var bayerContourNextUsb = require('../drivers/bayerContourNextUsb');
+var bayerContourNext = require('../drivers/bayerContourNext');
 
 var device = {
   log: require('../bows')('Device')
@@ -44,7 +44,7 @@ device._deviceDrivers = {
   'AsanteSNAP': asanteSnap,
   'InsuletOmniPod': insuletOmniPod,
   'OneTouchUltra2': oneTouchUltra2,
-  'BayerContourNextUsb': bayerContourNextUsb
+  'BayerContourNext': bayerContourNext
 };
 
 device._deviceComms = {
@@ -52,11 +52,15 @@ device._deviceComms = {
   'OneTouchMini': serialDevice,
   'AbbottFreeStyle': serialDevice,
   'AsanteSNAP': serialDevice,
-  'OneTouchUltra2': serialDevice
+  'OneTouchUltra2': serialDevice,
+  'BayerContourNext': bayerContourNext
 };
 
 device._silentComms = {};
 _.forEach(_.keys(device._deviceComms), function(driverId) {
+
+  console.info('device._silentComms.driverId',driverId);
+
   var comm = device._deviceComms[driverId];
   device._silentComms[driverId] = comm({silent: true});
   device._deviceComms[driverId] = comm();
@@ -88,6 +92,7 @@ device.init = function(options, cb) {
 
 device._getAllDriverManifests = function() {
   var manifest = chrome.runtime.getManifest();
+
   var usbDevicesPermission = _.find(manifest.permissions, function(permission) {
     return permission.usbDevices;
   });
@@ -198,6 +203,9 @@ device.detectHid = function(driverId, cb) {
 device.detectUsb = function(driverId, cb) {
   var self = this;
   var driverManifest = this.getDriverManifest(driverId);
+
+  console.info('el manifest', driverManifest);
+
   var identification = {
     vendorId: driverManifest.vendorId,
     productId: driverManifest.productId
@@ -244,6 +252,7 @@ device.detect = function(driverId, options, cb) {
     options = { version: self._version };
   }
   var driverManifest = this.getDriverManifest(driverId);
+  console.info('manifest del detect',driverManifest);
   this.detectUsb(driverId, function(err, devdata) {
     if (err) {
       return cb(err);
@@ -264,6 +273,9 @@ device.detect = function(driverId, options, cb) {
       return cb(null, devdata);
     });
   });
+
+
+
 };
 
 device.detectAll = function(cb) {
