+++ lib/core/deviceInfo.js
@@ -38,5 +38,5 @@ module.exports = {
                                  'Blood glucose and ketone meter'),
   'InsuletOmniPod': infoBuilder('Insulet OmniPod', 'Upload .ibf file from PDM.'),
   'OneTouchUltra2': infoBuilder('OneTouch Ultra2', ''),
-  'BayerContourNextUsb': infoBuilder('Bayer Contour', 'NEXT USB')
+  'BayerContourNext': infoBuilder('Bayer Contour Next', '')
 };
