+++ lib/driverManager.js
@@ -88,6 +88,11 @@ module.exports = function (driverObjects, configs) {
             if (err) {
               cb(err, result);
             } else {
+
+
+              console.info('result.model',result.model);
+              console.info('result.serialNumber',result.serialNumber);
+              console.info('result.id',result.id);
               cb(null, {
                 model: result.model,
                 serialNumber: result.serialNumber,
