+++ lib/drivers/bayerContourNext.js
