+++ lib/drivers/oneTouchMiniDriver.js
@@ -486,6 +486,9 @@ module.exports = function (config) {
 
     connect: function (progress, data, cb) {
       debug('in connect!');
+
+      data.deviceInfo.deviceId = data.id;
+
       cfg.deviceComms.connect(data.deviceInfo, oneTouchPacketHandler, probe, function(err) {
         if (err) {
           return cb(err);
