+++ lib/state/appState.js
@@ -61,9 +61,9 @@ appState.getInitial = function() {
       source: {type: 'device', driverId: 'OneTouchMini'}
     },
     {
-      name: 'BayerContourNextUSB',
-      key: 'bayercontournextusb',
-      source: {type: 'device', driverId: 'BayerContourNextUsb'}
+      name: 'BayerContourNext',
+      key: 'bayercontournext',
+      source: {type: 'device', driverId: 'BayerContourNext'}
     }
   ];
