+++ manifest.json
@@ -30,6 +30,7 @@
     "storage",
     "serial",
     "usb",
+    "hid",
     {
       "usbDevices": [
         {
@@ -86,8 +87,9 @@
           "sendTimeout": 5000
         },
         {
-          "deviceName": "BayerContourNextUsb",
-          "driverId": "BayerContourNextUsb",
+          "deviceName": "BayerContourNext",
+          "driverId": "BayerContourNext",
+          "mode": "HID",
           "vendorId": 6777,
           "productId": 29712
         }
